add separate pointer file, use xcursor
use xcursor because cursorfonts are deprecated and ugly. using xcursor
requires an xlib display, so this requires a shared xlib/xcb connection
(libX11-xcb). note that we could have used xcb-util-cursor instead, but
this is a less accessible dependency. also, the xlib display may come in
handy if we want to use xft.